src/main/java/com/xdavide9/demo/security/MyUserDetailsService.java
package com.xdavide9.demo.security;

import com.xdavide9.demo.jpa.repositories.user.UserRepository;
import lombok.AllArgsConstructor;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
@AllArgsConstructor
public class MyUserDetailsService implements UserDetailsService {

    private final UserRepository userRepository;

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        return userRepository.findByEmail(username)
                .map(UserDetailsDTO::new)
                .orElseThrow(() -> new UsernameNotFoundException(
                        String.format("User with email [%s] not found", username)
                ));
    }
}
